refactor: クイズの共通ロジックをuseColorQuizフックに抽出
ColorToHexとHexToColorの重複する状態管理・ハンドラをカスタムフックに統合する。
aria-labelをボタンに追加しStorybookのセレクタを堅牢化する。
