+++ apps/main/src/app/color-quiz/_components/color-quiz/hex-to-color.tsx
@@ -1,40 +1,21 @@
 'use client';
 
 import { Badge, Button } from '@k8o/arte-odyssey';
-import { type FC, useCallback, useMemo, useState } from 'react';
-import {
-  generateDistractors,
-  generateRandomHex,
-  getContrastTextColor,
-  shuffleArray,
-} from '../../_utils/color-quiz';
-
-type Phase = 'question' | 'result';
-
-const OPTION_COUNT = 4;
+import type { FC } from 'react';
+import { getContrastTextColor } from '../../_utils/color-quiz';
+import { useColorQuiz } from './use-color-quiz';
 
 export const HexToColor: FC = () => {
-  const [targetHex, setTargetHex] = useState(() => generateRandomHex());
-  const [selectedHex, setSelectedHex] = useState<string | null>(null);
-  const [phase, setPhase] = useState<Phase>('question');
-
-  const options = useMemo(() => {
-    const distractors = generateDistractors(targetHex, OPTION_COUNT - 1);
-    return shuffleArray([targetHex, ...distractors]);
-  }, [targetHex]);
-
-  const handleSubmit = useCallback(() => {
-    if (selectedHex === null) return;
-    setPhase('result');
-  }, [selectedHex]);
-
-  const handleNext = useCallback(() => {
-    setTargetHex(generateRandomHex());
-    setSelectedHex(null);
-    setPhase('question');
-  }, []);
-
-  const isCorrect = selectedHex === targetHex;
+  const {
+    targetHex,
+    selectedHex,
+    setSelectedHex,
+    phase,
+    options,
+    isCorrect,
+    handleSubmit,
+    handleNext,
+  } = useColorQuiz();
 
   return (
     <div className="flex flex-col gap-6">
